Audit Report Bug Fix - CommitmentGenerator - scaledAmount size mismatch scaledAmount size should be uint120
Changes are as follows -
- Changed the size of scaledAmount from uint96 to uint120

+++ panther-protocol-contracts-ws/contracts/pantherPool/CommitmentGenerator.sol
@@ -17,7 +17,7 @@ abstract contract CommitmentGenerator {
     function generateCommitment(
         uint256 pubSpendingKeyX,
         uint256 pubSpendingKeyY,
-        uint96 scaledAmount,
+        uint120 scaledAmount,
         uint160 zAssetId,
         uint32 creationTime
     ) internal pure returns (bytes32 commitment) {
